requirements.txt
aiodns==3.0.0
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
emoji==1.7.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pycares==4.1.2
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
ujson==5.2.0
uvloop==0.16.0
yarl==1.7.2
